chore(editor): rename SignatureCanvas to signature-canvas for kebab-case lint rule
