+++ packages/editor/src/components/signature-view/signature-canvas.tsx
